Detect locale earlier to remove test warnings
