+++ src/base/services/i18n.js
@@ -8,6 +8,7 @@ import locales, { messages as loadMessages, quasarMessages as loadQuasarMessages
 import { defineService } from '@/utils/datastore/helpers'
 import dateFnsHelper from '@/utils/dateFnsHelper'
 import polyfill from '@/utils/polyfill'
+import { detectLocale } from '@/utils/utils'
 
 export const DEFAULT_LOCALE = 'en'
 
@@ -62,39 +63,3 @@ export const useI18nService = defineService(() => {
     setLocale,
   }
 })
-
-export function detectLocale () {
-  let requested = []
-  if (navigator.languages) {
-    navigator.languages.forEach(e => {
-      requested.push(e.toLowerCase())
-      // detect similar languages with slightly less priority
-      if (e.includes('-')) {
-        requested.push(e.replace(/-.*$/, '').toLowerCase())
-      }
-      // alias definitions
-      else if (e === 'zh') {
-        requested.push('zh-hans', 'zh-hant')
-      }
-      else if (e === 'zh_TW') {
-        requested.push('zh-hant')
-      }
-      else if (e === 'zh_CN') {
-        requested.push('zh-hans')
-      }
-    })
-  }
-  else {
-    const val =
-      navigator.language ||
-      navigator.browserLanguage ||
-      navigator.systemLanguage ||
-      navigator.userLanguage
-    if (val) {
-      requested = [val.toLowerCase()]
-    }
-  }
-  if (requested) {
-    return requested.find(e => locales[e])
-  }
-}
